requirements.txt
Django==1.7.4
Pillow==2.6.1
django-allauth==0.18.0
django-crispy-forms==1.4.0
django-filetransfers==0.1.0
oauthlib==0.6.3
python-dateutil==2.2
python-mimeparse==0.1.4
python-openid==2.2.5
requests==2.4.3
requests-oauthlib==0.4.1
six==1.8.0
stripe==1.27.1
wsgiref==0.1.2
